action method from the Profile model removed
I removed all action methods in the Profile model, as they were violating the SOLID principle of Single responsibility.
Perhaps it would be necessary to just segregate these methods in other classes, however, it is not necessary, since these methods will only be used once in my application (in each view) and have little logic. Segregating these methods in practice will not bring any benefit to the maintenance or refactoring of my code